move idVersion to BaseElement
