+++ celements-xwiki-core/src/main/java/com/xpn/xwiki/objects/BaseElement.java
@@ -21,6 +21,7 @@
 
 package com.xpn.xwiki.objects;
 
+import com.celements.store.id.IdVersion;
 import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.XWikiException;
 import com.xpn.xwiki.doc.XWikiDocument;
@@ -69,6 +70,19 @@ public abstract class BaseElement implements ElementInterface, Serializable
      */
     private EntityReferenceSerializer<String> localEntityReferenceSerializer =
         Utils.getComponent(EntityReferenceSerializer.class, "local");
+
+    private IdVersion idVersion = IdVersion.XWIKI_2;
+
+    public IdVersion getIdVersion() {
+      verifyIdVersion();
+      return idVersion;
+    }
+    
+    private void verifyIdVersion() {
+      if (idVersion == null) {
+        throw new IllegalStateException("no id version set");
+      }
+    }
 
     /**
      * {@inheritDoc}
